# Conflicts:
#	src/main/java/com/roadscanner/domain/upload/FileUploadVO.java
#	src/main/java/com/roadscanner/service/upload/FileUploadServiceImpl.java
#	src/main/resources/mapper/upload/upload.xml
#	src/main/webapp/WEB-INF/root-context.xml
#	src/main/webapp/WEB-INF/views/login/admin.jsp
#	src/main/webapp/WEB-INF/views/login/mypage.jsp
#	src/main/webapp/WEB-INF/views/login/navbar.jsp

+++ src/main/java/com/roadscanner/domain/upload/FileUploadVO.java
@@ -12,7 +12,7 @@ public class FileUploadVO extends DTO {
 	private String uploadDate; // 등록일
 	private String name; // 저장파일명 : YYMMDDHH24MISS_원본파일명
 	private String url; // URL : S3 저장 링크
-	private long fileSize; // 파일크기
+	private double fileSize; // 파일크기
 	private int checked; // 검토여부
 	private int u1; // 싫어요 이유1 (다른 모양)
 	private int u2; // 싫어요 이유2 (다른 색깔)
@@ -22,7 +22,8 @@ public class FileUploadVO extends DTO {
 	public FileUploadVO() {
 	}
 
-	public FileUploadVO(int idx, String id, int category, String uploadDate, String name, String url, long fileSize,
+
+	public FileUploadVO(int idx, String id, int category, String uploadDate, String name, String url, double fileSize,
 			int checked, int u1, int u2, int u3) {
 		super();
 		this.idx = idx;
@@ -86,11 +87,11 @@ public class FileUploadVO extends DTO {
 		this.url = url;
 	}
 
-	public long getFileSize() {
+	public double getFileSize() {
 		return fileSize;
 	}
 
-	public void setFileSize(long fileSize) {
+	public void setFileSize(double fileSize) {
 		this.fileSize = fileSize;
 	}
